+++ CHANGES.md
@@ -1,3 +1,8 @@
+
+
+- The `mli`, `cmi`, `cmo`, `cmti`, `cmx` and `cmt` commands are grouped in
+  the `cobjs` command.
+
 v0.0.1 2016-09-23 Zagreb
 ------------------------